refactor: package structure
