+++ minibots/__init__.py
@@ -1,0 +1,8 @@
+from minibots.bot import Bot
+from minibots.IRepository import IBot
+from minibots.entity import entity
+from minibots.bots.Github.ghrepository import ghrepository
+from minibots.bots.Github.ghuser import ghuser
+from minibots.bots.Github.githubbot import GithubBot
+from minibots.bots.Github.githubrepository import GithubRepository
+from minibots.bots.Github.ghrequester import requestsGithub
