+++ minibots/bots/Github/ghrequester.py
@@ -1,5 +1,4 @@
 from requests import request, Response
-from bots.Github import common
 
 class requestsGithub():
