+++ minibots/bots/Github/ghuser.py
@@ -1,8 +1,7 @@
-import common
-from bot.entity import entity
+from minibots import entity
 import json
 
-class user(entity):
+class ghuser(entity):
     
     username: str
     url: str
