+++ minibots/test/bots/Github/test_githubbot.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch
-from bots.Github import common 
-from bots.Github.githubrepository import GithubRepository
+import common 
+from minibots import GithubRepository
 
 class TestGithubBot(unittest.TestCase):
 
@@ -9,7 +9,7 @@ class TestGithubBot(unittest.TestCase):
 
     def test_getUser(self):
 
-        with patch('bots.Github.githubrepository.requests.get') as mock_get:
+        with patch('minibots.bots.Github.githubrepository.requests.get') as mock_get:
             mock_get.return_value.status_code=200
             obj = GithubRepository()
             response = obj.bconnect(self.resource)
