--- test/bots/Github/common.py
@@ -1,4 +1,0 @@
-import os, sys
-currentdir = os.path.dirname(os.path.realpath(__file__))
-parentdir = os.path.dirname(os.path.dirname(os.path.dirname(currentdir)))
-sys.path.append(parentdir)
